Use bash glob patterns instead of "grep -P" for compatibility with macOS
On macOS, grep doesn't support Perl regular expressions (-P), which we used
for the tests, in order to match the NUL character after replacing newlines,
so that we could match against the whole output and not only line by line.

After all it seems somwhat simpler and more portable to use bash and compare
the output against a glob pattern.